requirements.txt
et-xmlfile>=1.1.0
numpy>=1.22.0
openpyxl>=3.0.9
pandas>=1.3.5
python-dateutil>=2.8.2
pytz>=2021.3
six>=1.16.0
olefile>=0.46

